refactor: moved interfaces/enums from muc plugin to dedicated files
