+++ demo/app/multi-user-chat/multi-user-chat.component.html
@@ -24,7 +24,7 @@
         <button (click)="subscribeWithMucSub(occupantJid)">subscribe with MUC/Sub</button>
         <button (click)="unsubscribeFromMucSub(occupantJid)">unsubscribe from MUC/Sub</button>
         <button (click)="getSubscriptions()">get MUC/Sub subscriptions</button>
-        <button (click)="queryMemberList(occupantJid)">get member list</button>
+        <button (click)="queryUserList(occupantJid)">get user list</button>
 
         <div>Current MUC subscriptions to room: {{occupantJid}}</div>
         <ul>
