+++ src/lib/components/chat-window/chat-window.component.ts
@@ -2,13 +2,13 @@ import { Component, Inject, Input, OnDestroy, OnInit, Optional, ViewChild } from
 import { Observable, Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
 import { Direction, Message } from '../../core/message';
-import { RoomMessage } from '../../services/adapters/xmpp/plugins/multi-user-chat.plugin';
 import { ChatContactClickHandler, CONTACT_CLICK_HANDLER_TOKEN } from '../../hooks/chat-contact-click-handler';
 import { ChatListStateService, ChatWindowState } from '../../services/chat-list-state.service';
 import { CHAT_SERVICE_TOKEN, ChatService } from '../../services/chat-service';
 import { ChatMessageInputComponent } from '../chat-message-input/chat-message-input.component';
 import { ChatMessageListComponent } from '../chat-message-list/chat-message-list.component';
 import {FILE_UPLOAD_HANDLER_TOKEN, FileUploadHandler} from '../../hooks/file-upload-handler';
+import { RoomMessage } from '../../services/adapters/xmpp/plugins/multi-user-chat/room.message';
 
 @Component({
     selector: 'ngx-chat-window',
