+++ src/lib/services/adapters/xmpp/plugins/message-archive.plugin.ts
@@ -7,7 +7,7 @@ import { Stanza } from '../../../../core/stanza';
 import { LogService } from '../../../log.service';
 import { XmppChatAdapter } from '../xmpp-chat-adapter.service';
 import { AbstractXmppPlugin } from './abstract-xmpp-plugin';
-import { MultiUserChatPlugin } from './multi-user-chat.plugin';
+import { MultiUserChatPlugin } from './multi-user-chat/multi-user-chat.plugin';
 import { ServiceDiscoveryPlugin } from './service-discovery.plugin';
 import { PUBSUB_EVENT_XMLNS } from './publish-subscribe.plugin';
 import { MessagePlugin } from './message.plugin';
