+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/affiliation.modification.ts
@@ -1,0 +1,8 @@
+import { JID } from '@xmpp/jid';
+import { Affiliation } from './affiliation';
+
+export interface AffiliationModification {
+    userJid: JID;
+    affiliation: Affiliation;
+    reason?: string;
+}
