+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/affiliation.ts
@@ -1,0 +1,7 @@
+export enum Affiliation {
+    none = 'none',
+    outcast = 'outcast',
+    member = 'member',
+    admin = 'admin',
+    owner = 'owner',
+}
