+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/invitation.ts
@@ -1,0 +1,9 @@
+import { JID } from '@xmpp/jid';
+
+export interface Invitation {
+    type: 'invite' | 'decline';
+    roomJid: JID;
+    roomPassword?: string;
+    from: JID;
+    message?: string;
+}
