+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/occupant.change.ts
@@ -1,0 +1,9 @@
+import { RoomOccupant } from './room.occupant';
+
+export type OccupantChangeType = 'joined' | 'left' | 'kicked' | 'banned' | 'changedNick' | 'invitation';
+
+export interface OccupantChange {
+    occupant: RoomOccupant;
+    change: OccupantChangeType;
+    newNick?: string;
+}
