+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/role.modification.ts
@@ -1,0 +1,7 @@
+import { Role } from './role';
+
+export interface RoleModification {
+    nick: string;
+    role: Role;
+    reason?: string;
+}
