+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/role.ts
@@ -1,0 +1,6 @@
+export enum Role {
+    none = 'none',
+    visitor = 'visitor',
+    participant = 'participant',
+    moderator = 'moderator',
+}
