+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/room.message.ts
@@ -1,0 +1,6 @@
+import { Message } from '../../../../../core/message';
+import { JID } from '@xmpp/jid';
+
+export interface RoomMessage extends Message {
+    from: JID;
+}
