+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/room.occupant.ts
@@ -1,0 +1,10 @@
+import { JID } from '@xmpp/jid';
+import { Affiliation } from './affiliation';
+import { Role } from './role';
+
+export interface RoomOccupant {
+    occupantJid: JID;
+    affiliation: Affiliation;
+    nick: string;
+    role: Role;
+}
