+++ src/lib/services/adapters/xmpp/plugins/multi-user-chat/room.user.ts
@@ -1,0 +1,12 @@
+import { JID } from '@xmpp/jid';
+import { Affiliation } from './affiliation';
+import { Role } from './role';
+
+export interface RoomUser {
+    userIdentifiers: {
+        userJid: JID,
+        nick?: string
+    }[];
+    affiliation?: Affiliation;
+    role?: Role;
+}
