+++ src/lib/services/adapters/xmpp/xmpp-chat-adapter.service.ts
@@ -16,7 +16,7 @@ import { ContactFactoryService } from '../../contact-factory.service';
 import { LogService } from '../../log.service';
 import { MessageArchivePlugin } from './plugins/message-archive.plugin';
 import { MessagePlugin } from './plugins/message.plugin';
-import { MultiUserChatPlugin } from './plugins/multi-user-chat.plugin';
+import { MultiUserChatPlugin } from './plugins/multi-user-chat/multi-user-chat.plugin';
 import { RosterPlugin } from './plugins/roster.plugin';
 import { XmppChatConnectionService, XmppChatStates } from './xmpp-chat-connection.service';
