+++ src/public-api.ts
@@ -37,7 +37,7 @@ export * from './lib/services/adapters/xmpp/plugins/message-state.plugin';
 export * from './lib/services/adapters/xmpp/plugins/message-uuid.plugin';
 export * from './lib/services/adapters/xmpp/plugins/message.plugin';
 export * from './lib/services/adapters/xmpp/plugins/muc-sub.plugin';
-export * from './lib/services/adapters/xmpp/plugins/multi-user-chat.plugin';
+export * from './lib/services/adapters/xmpp/plugins/multi-user-chat/multi-user-chat.plugin';
 export * from './lib/services/adapters/xmpp/plugins/ping.plugin';
 export * from './lib/services/adapters/xmpp/plugins/publish-subscribe.plugin';
 export * from './lib/services/adapters/xmpp/plugins/push.plugin';
@@ -59,3 +59,13 @@ export { LinkOpener, LINK_OPENER_TOKEN } from './lib/components/chat-message-lin
 export { selectFile } from './lib/core/utils-file';
 export { jid as parseJid } from '@xmpp/client';
 export { JID } from '@xmpp/jid';
+export { Room } from './lib/services/adapters/xmpp/plugins/multi-user-chat/room';
+export { Affiliation } from './lib/services/adapters/xmpp/plugins/multi-user-chat/affiliation';
+export { Role } from './lib/services/adapters/xmpp/plugins/multi-user-chat/role';
+export { RoomUser } from './lib/services/adapters/xmpp/plugins/multi-user-chat/room.user';
+export { RoomOccupant } from './lib/services/adapters/xmpp/plugins/multi-user-chat/room.occupant';
+export { Invitation } from './lib/services/adapters/xmpp/plugins/multi-user-chat/invitation';
+export { RoomMessage } from './lib/services/adapters/xmpp/plugins/multi-user-chat/room.message';
+export { RoleModification } from './lib/services/adapters/xmpp/plugins/multi-user-chat/role.modification';
+export { OccupantChange } from './lib/services/adapters/xmpp/plugins/multi-user-chat/occupant.change';
+export { AffiliationModification } from './lib/services/adapters/xmpp/plugins/multi-user-chat/affiliation.modification';
